Make YamlConfiguration field final
The implementation is the same as YamlConfiguration#loadConfiguration(File) though we manually load it in instead of the factory method earlier mentioned. This increases immutability. The only difference from the factory method is that we now throw a RuntimeException when a FileNotFoundException is thrown. We throw a RuntimeException firmly to mimic the behavior of how ConfigurateConfigAdapter#reload() handles exceptions.