Add modifyTrack to TrackManager
The UserManager and GroupManager both have a convenience method for
loading and saving a user or a group after applying a modification. The
TrackManager lacks this operation, despite also being modifiable and
needing to be saved. I see this as a small hole in the API.

This PR adds a modifyTrack method to TrackManager, very closely modeled
after modifyGroup.

I am making this PR because I found myself wanting this method, and it
provides symmetry in the API.