Fix QueryOptionsCache referencing outdated player instance
Respawning creates a new ServerPlayerEntity instance. The current code
copies the QueryOptionsCache from the old player instance, however this
causes an issue, because the cache still holds the reference to the old
entity. So any usage of it will invoke ContextManager.calculate causing
outdated context to be calculated. This is simply fixed by creating a
new QueryOptionsCache with the new instance.